fix: refactor app router setup and correct import path for AppRouter
