+++ src/trpc/routers/_app.ts
@@ -1,8 +1,10 @@
 import { agentsRouter } from '@/modules/agents/server/procedures';
 import { createContext } from '@/trpc/context';
-import { createCallerFactory, mergeRouters } from '@/trpc/trpc';
+import { createCallerFactory, router } from '@/trpc/trpc';
 
-export const appRouter = mergeRouters(agentsRouter);
+export const appRouter = router({
+  agents: agentsRouter,
+});
 export const createCaller = createCallerFactory(appRouter);
 
 export const createAsyncCaller = async () => {
