+++ src/trpc/server.tsx
@@ -4,8 +4,8 @@ import { createTRPCOptionsProxy } from '@trpc/tanstack-react-query';
 import { cache } from 'react';
 
 import { createContext } from './context';
-import { appRouter } from './middleware/routers/_app';
 import { makeQueryClient } from './query-client';
+import { appRouter } from './routers/_app';
 
 export const getQueryClient = cache(makeQueryClient);
