+++ src/utils/trpc.ts
@@ -1,6 +1,6 @@
 import { createTRPCContext } from '@trpc/tanstack-react-query';
 
-import type { AppRouter } from '@/trpc/middleware/routers/_app';
+import type { AppRouter } from '@/trpc/routers/_app';
 
 export const { TRPCProvider, useTRPC, useTRPCClient } =
   createTRPCContext<AppRouter>();
